Move class ParticleBase to own file
